Refactor lots of things
