+++ src/app/webhooks/[webhookId]/edit/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+const EditWebhookPage = () => {
+  return <p>Edit webhook page will go here</p>;
+};
+
+export default EditWebhookPage;
