+++ src/components/Webhooks/index.ts
@@ -1,1 +1,6 @@
 export * from './WebhooksPage';
+export { WebhookDescriptionField } from '@/components/Webhooks/fields/Description';
+export { WebhookIntervalField } from '@/components/Webhooks/fields/Interval';
+export { WebhookServiceField } from '@/components/Webhooks/fields/Service';
+export { WebhookTargetUriField } from '@/components/Webhooks/fields/TargetURI';
+export { WebhookVerificationTokenField } from '@/components/Webhooks/fields/VerificationToken';
