--- src/components/Webhooks/steps/Configuration/Configuration.tsx
@@ -1,73 +1,0 @@
-import React, { FC } from 'react';
-import { useFormContext } from 'react-hook-form';
-import { Label } from '@/components/Label';
-import { TextField } from '@/components/TextField';
-import { TextError } from '@/components/TextError';
-import { SelectField } from '@/components/SelectField';
-import { WebhookFormInput } from '@/types/webhook';
-import { CELBuilder } from './CELBuilder';
-
-export const WebhookConfigStep: FC = () => {
-  const {
-    register,
-    control,
-    formState: { errors },
-    getValues,
-  } = useFormContext<WebhookFormInput>();
-  return (
-    <>
-      <div className={'flex flex-col gap-2.5'}>
-        <Label>Webhook description</Label>
-        <TextField
-          {...register('description', {
-            required: 'Please enter a webhook description',
-          })}
-          placeholder="Enter a description for your webhook"
-        />
-        {errors.description && (
-          <TextError
-            errorMessage={
-              typeof errors.description.message === 'string'
-                ? errors.description.message
-                : ''
-            }
-          />
-        )}
-      </div>
-      <div className={'flex flex-col gap-2.5'}>
-        <Label>Service</Label>
-        <SelectField
-          value={getValues('service')}
-          placeholder={'Choose a service'}
-          {...register('service', {
-            required: 'Please select a webhook service',
-          })}
-          options={[{ value: 'Telemetry', text: 'Telemetry' }]}
-          control={control}
-        />
-        {errors.service && (
-          <TextError errorMessage={(errors.service.message as string) ?? ''} />
-        )}
-      </div>
-      <CELBuilder />
-      <div className={'flex flex-col gap-2.5'}>
-        <Label>Interval</Label>
-        <SelectField
-          value={getValues('setup')}
-          placeholder={'Choose an interval'}
-          {...register('setup', {
-            required: 'Please select a webhook interval',
-          })}
-          options={[
-            { value: 'Realtime', text: 'Realtime' },
-            { value: 'Hourly', text: 'Hourly' },
-          ]}
-          control={control}
-        />
-        {errors.setup && (
-          <TextError errorMessage={(errors.setup.message as string) ?? ''} />
-        )}
-      </div>
-    </>
-  );
-};
