+++ src/config/navigation.ts
@@ -11,6 +11,7 @@ const APP_DETAILS_REGEX = /^\/app\/details\/[^/]+$/;
 const LICENSE_DETAILS_REGEX = /^\/license\/details\/[^/]+$/;
 const LICENSED_VEHICLES_REGEX = /^\/license\/vehicles\/[^/]+$/;
 const CREATE_WEBHOOK_REGEX = /^\/webhooks\/create\/[^/]+$/;
+const EDIT_WEBHOOK_REGEX = /^\/webhooks\/[^/]+\/edit$/;
 
 export const getPageTitle = (path: string) => {
   const staticPageTitle = pageTitles[path];
@@ -19,6 +20,7 @@ export const getPageTitle = (path: string) => {
   if (LICENSE_DETAILS_REGEX.test(path)) return 'License Details';
   if (LICENSED_VEHICLES_REGEX.test(path)) return 'Licensed Vehicles';
   if (CREATE_WEBHOOK_REGEX.test(path)) return 'Create a webhook';
+  if (EDIT_WEBHOOK_REGEX.test(path)) return 'Edit webhook';
 };
 
 const pageTitles: Record<string, string> = {
