--- src/hooks/useWebhooks.ts
@@ -1,190 +1,0 @@
-import { useEffect, useState } from 'react';
-import {
-  createWebhook,
-  deleteWebhook,
-  fetchSignalNames,
-  fetchWebhooks,
-  generateCEL,
-  updateWebhook,
-  fetchWebhookById,
-} from '@/services/webhook';
-import { Condition, Webhook } from '@/types/webhook';
-import { useQuery } from '@tanstack/react-query';
-
-import { getDevJwt } from '@/utils/devJwt';
-
-const handleFetchWebhooks = async (clientId: string) => {
-  const token = getDevJwt(clientId);
-  if (!token) throw new Error(`No devJWT found for client ${clientId}`);
-  return await fetchWebhooks({ token });
-};
-
-export const useWebhooks = (clientId: string, options?: { enabled?: boolean }) => {
-  return useQuery({
-    queryKey: ['webhooks', clientId],
-    queryFn: () => handleFetchWebhooks(clientId),
-    ...options,
-  });
-};
-
-interface FetchWebhookVehiclesByIdParams {
-  webhookId: string;
-  clientId: string;
-}
-const handleFetchWebhookVehiclesById = async ({
-  webhookId,
-  clientId,
-}: FetchWebhookVehiclesByIdParams) => {
-  const token = getDevJwt(clientId);
-  if (!token) throw new Error(`No devJWT found for client ${clientId}`);
-  return await fetchWebhookById({ webhookId, token });
-};
-
-export const useWebhookVehiclesById = (params: FetchWebhookVehiclesByIdParams) => {
-  return useQuery({
-    queryKey: ['webhook-vehicles', params],
-    queryFn: () => handleFetchWebhookVehiclesById(params),
-  });
-};
-
-export const DEPRECATED_useWebhooks = () => {
-  const [webhooks] = useState<Webhook[]>([]);
-  const [currentWebhook, setCurrentWebhook] = useState<Partial<Webhook> | null>(null);
-  const [parametersInput, setParametersInput] = useState<string>('{}');
-  const [showDeleteConfirm, setShowDeleteConfirm] = useState(false);
-  const [webhookToDelete, setWebhookToDelete] = useState<Webhook | null>(null);
-  const [signalNames, setSignalNames] = useState<string[]>([]);
-  const [conditions, setConditions] = useState<Condition[]>([]);
-  const [logic, setLogic] = useState('AND');
-  const [generatedCEL, setGeneratedCEL] = useState('');
-  const [expandedWebhook, setExpandedWebhook] = useState<string | null>(null);
-
-  const loadWebhooks = async () => {
-    try {
-      // const data = await fetchWebhooks();
-      // setWebhooks(data);
-    } catch (error) {
-      console.error('Error fetching webhooks:', error);
-    }
-  };
-
-  const loadSignalNames = async () => {
-    try {
-      const signals = await fetchSignalNames();
-      setSignalNames(signals);
-    } catch (error) {
-      console.error('Error fetching signal names:', error);
-    }
-  };
-
-  useEffect(() => {
-    loadWebhooks();
-    loadSignalNames();
-  }, []);
-
-  useEffect(() => {
-    const generateAndSetCEL = async () => {
-      try {
-        if (conditions.length > 0 && logic) {
-          const celExpression = await generateCEL({ conditions, logic });
-          setGeneratedCEL(celExpression);
-          if (currentWebhook) {
-            setCurrentWebhook({
-              ...currentWebhook,
-              trigger: celExpression,
-            });
-          }
-        }
-      } catch (error) {
-        console.error('Failed to generate CEL:', error);
-      }
-    };
-
-    generateAndSetCEL();
-  }, [logic, conditions]);
-
-  const handleCreate = async () => {
-    try {
-      const parsedParameters = JSON.parse(parametersInput);
-      const payload = {
-        ...currentWebhook,
-        trigger: generatedCEL,
-        parameters: parsedParameters,
-      };
-      await createWebhook(payload as Webhook);
-      await loadWebhooks();
-      resetForm();
-    } catch (error) {
-      console.error('Error creating webhook:', error);
-      alert('Invalid JSON in Parameters field.');
-    }
-  };
-
-  const handleUpdate = async () => {
-    if (!currentWebhook?.id) return;
-    try {
-      const parsedParameters = JSON.parse(parametersInput);
-      const payload = {
-        ...currentWebhook,
-        trigger: generatedCEL,
-        parameters: parsedParameters,
-      };
-      await updateWebhook(currentWebhook.id, payload, '');
-      await loadWebhooks();
-      resetForm();
-    } catch (error) {
-      console.error('Error updating webhook:', error);
-      alert('Invalid JSON in Parameters field.');
-    }
-  };
-
-  const handleDelete = async () => {
-    if (webhookToDelete) {
-      await deleteWebhook(webhookToDelete.id);
-      await loadWebhooks();
-      setWebhookToDelete(null);
-      setShowDeleteConfirm(false);
-    }
-  };
-
-  const resetForm = () => {
-    setCurrentWebhook(null);
-    setParametersInput('{}');
-    setConditions([]);
-    setGeneratedCEL('');
-    setLogic('AND');
-  };
-
-  const handleShowCreateForm = () => {
-    resetForm();
-    setCurrentWebhook({
-      description: 'My Cool Webhook',
-      target_uri: 'https://example.com/webhook',
-    });
-  };
-
-  return {
-    webhooks,
-    currentWebhook,
-    setCurrentWebhook,
-    parametersInput,
-    setParametersInput,
-    conditions,
-    setConditions,
-    logic,
-    setLogic,
-    signalNames,
-    generatedCEL,
-    expandedWebhook,
-    setExpandedWebhook,
-    showDeleteConfirm,
-    setShowDeleteConfirm,
-    webhookToDelete,
-    setWebhookToDelete,
-    handleCreate,
-    handleUpdate,
-    handleDelete,
-    resetForm,
-    handleShowCreateForm,
-  };
-};
